requirements.txt
Flask==1.1.1
line-bot-sdk==1.15.0
boto3==1.9.237
pya3rt==1.1